@guilhermef 2016-01-16
Thanks @aectann, that's an awesome change.
I'll release a new version.